# copr: use criterion-perf-events to get instructions