--- a/tests/benches/deadlock_detector/mod.rs
+++ b/tests/benches/deadlock_detector/mod.rs
@@ -76,7 +76,6 @@ fn bench_dense_detect_without_cleanup(c: &mut Criterion) {
         100,
         1_000,
         10_000,
-        10_000,
         100_000,
         1_000_000,
         10_000_000,