--- a/tests/benches/hierarchy/engine/mod.rs
+++ b/tests/benches/hierarchy/engine/mod.rs
@@ -79,11 +79,19 @@ fn bench_engine_get<E: Engine, F: EngineFactory<E>>(
 }
 
 pub fn bench_engine<E: Engine, F: EngineFactory<E>>(c: &mut Criterion, configs: &[BenchConfig<F>]) {
-    c.bench_function_over_inputs(
-        "engine_get(exclude snapshot)",
-        bench_engine_get,
-        configs.to_vec(),
-    );
-    c.bench_function_over_inputs("engine_put", bench_engine_put, configs.to_owned());
-    c.bench_function_over_inputs("engine_snapshot", bench_engine_snapshot, configs.to_owned());
+    let mut group = c.benchmark_group("engine");
+    for config in configs {
+        group.bench_with_input(
+            format!("get(exclude snapshot)/{:?}", config),
+            config,
+            bench_engine_get,
+        );
+        group.bench_with_input(format!("put/{:?}", config), config, bench_engine_put);
+        group.bench_with_input(
+            format!("snapshot/{:?}", config),
+            config,
+            bench_engine_snapshot,
+        );
+    }
+    group.finish();
 }