--- a/tests/benches/hierarchy/txn/mod.rs
+++ b/tests/benches/hierarchy/txn/mod.rs
@@ -178,21 +178,25 @@ fn txn_rollback_non_prewrote<E: Engine, F: EngineFactory<E>>(
 }
 
 pub fn bench_txn<E: Engine, F: EngineFactory<E>>(c: &mut Criterion, configs: &[BenchConfig<F>]) {
-    c.bench_function_over_inputs("txn_prewrite", txn_prewrite, configs.to_owned());
-    c.bench_function_over_inputs("txn_commit", txn_commit, configs.to_owned());
-    c.bench_function_over_inputs(
-        "txn_rollback_prewrote",
-        txn_rollback_prewrote,
-        configs.to_owned(),
-    );
-    c.bench_function_over_inputs(
-        "txn_rollback_conflict",
-        txn_rollback_conflict,
-        configs.to_owned(),
-    );
-    c.bench_function_over_inputs(
-        "txn_rollback_non_prewrote",
-        txn_rollback_non_prewrote,
-        configs.to_owned(),
-    );
+    let mut group = c.benchmark_group("txn");
+    for config in configs {
+        group.bench_with_input(format!("prewrite/{:?}", config), config, txn_prewrite);
+        group.bench_with_input(format!("commit/{:?}", config), config, txn_commit);
+        group.bench_with_input(
+            format!("rollback_prewrote/{:?}", config),
+            config,
+            txn_rollback_prewrote,
+        );
+        group.bench_with_input(
+            format!("rollback_conflict/{:?}", config),
+            config,
+            txn_rollback_conflict,
+        );
+        group.bench_with_input(
+            format!("rollback_non_prewrote/{:?}", config),
+            config,
+            txn_rollback_non_prewrote,
+        );
+    }
+    group.finish();
 }