Signed-off-by: Yang Keao <keao.yang@yahoo.com>

### What problem does this PR solve?

Remove `PAPI` dependency, which is really out of fashion now :face_with_head_bandage: , as the perf event document gets richer.

### What is changed and how it works?

Use `criterion-perf-events` rather than `criterion-papi` to get instructions counts during the benchmark.

The only degradation is that the unit is displayed as `cycles`, which is hard coded in `criterion-perf-events` :crying_cat_face: . Don't mind it because it doesn't effect the correctness of the result.

```
table_scan_datum_absent_large_row/RocksDB/batch/with_dag/rows=5000                                                                            
                        time:   [117543777.9983 cycles 117816414.1483 cycles 118078450.0667 cycles]

table_scan_long_datum_all/Memory/batch/next=1024                                                                            
                        time:   [1839019.2600 cycles 1839588.7667 cycles 1840377.1964 cycles]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
table_scan_long_datum_all/RocksDB/normal/with_dag/rows=5000                                                                             
                        time:   [20458704.2422 cycles 20462095.8382 cycles 20466490.3271 cycles]
Found 17 outliers among 100 measurements (17.00%)
  13 (13.00%) low severe
  4 (4.00%) high severe
table_scan_long_datum_all/RocksDB/batch/with_dag/rows=5000                                                                             
                        time:   [20461674.2794 cycles 20463698.1418 cycles 20466326.1559 cycles]
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low severe
  4 (4.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe
```

Besides this change, this PR also:

1. upgrade criterion to 0.3.4 and remove deprecated function calls
2. fix #9671

### Release note <!-- bugfixes or new feature need a release note -->

* No release note