@Xuanwo 2021-02-19
```
[2021-02-19T10:34:51.347Z] error: use of deprecated associated function `criterion::Criterion::<M>::bench_function_over_inputs`: Please use BenchmarkGroups instead.

[2021-02-19T10:34:51.347Z]   --> tests/benches/deadlock_detector/mod.rs:91:7

[2021-02-19T10:34:51.347Z]    |

[2021-02-19T10:34:51.347Z] 91 |     c.bench_function_over_inputs("bench_dense_detect_without_cleanup", bench_detect, cfgs);

[2021-02-19T10:34:51.347Z]    |       ^^^^^^^^^^^^^^^^^^^^^^^^^^

[2021-02-19T10:34:51.347Z]    |

```

Maybe we can solve this warning?