@YangKeao 2021-02-19
@Xuanwo Yes. These functions are deprecated on `criterion 0.3.4`, and they are removed now (in this PR).