@sticnarf 2021-02-19
> Oops, found a bug for one benchmark while testing this PR smiley_cat
> 
> #9671, issue has been submitted.

Could you help fix it in this PR? It is easy to fix :)

```
diff --git a/tests/benches/hierarchy/mvcc/mod.rs b/tests/benches/hierarchy/mvcc/mod.rs
index 47ba221e9..28caa111a 100644
--- a/tests/benches/hierarchy/mvcc/mod.rs
+++ b/tests/benches/hierarchy/mvcc/mod.rs
@@ -36,7 +36,7 @@ where
     .generate(DEFAULT_ITERATIONS);
     for (k, v) in &kvs {
         let txn_props = TransactionProperties {
-            start_ts: TimeStamp::default(),
+            start_ts,
             kind: TransactionKind::Optimistic(false),
             commit_kind: CommitKind::TwoPc,
             primary: &k.clone(),
diff --git a/tests/benches/hierarchy/txn/mod.rs b/tests/benches/hierarchy/txn/mod.rs
index 6fe7574b1..5efee71a3 100644
--- a/tests/benches/hierarchy/txn/mod.rs
+++ b/tests/benches/hierarchy/txn/mod.rs
@@ -32,7 +32,7 @@ where
     let kvs = KvGenerator::new(config.key_length, config.value_length).generate(DEFAULT_ITERATIONS);
     for (k, v) in &kvs {
         let txn_props = TransactionProperties {
-            start_ts: TimeStamp::default(),
+            start_ts,
             kind: TransactionKind::Optimistic(false),
             commit_kind: CommitKind::TwoPc,
             primary: &k.clone(),
```